Move CLI command to /shoutrrr
This will make the CLI command installable by running:

```
go install github.com/containrrr/shoutrrr/shoutrrr@latest
```

When installing using `go install` previously the command got installed
with the binary name `cli`.
