+++ build.sh
@@ -1,3 +1,3 @@
 #!/bin/sh
 
-go build -o ./shoutrrr ./cli
+go build -o ./shoutrrr/shoutrrr ./shoutrrr
