+++ generate-service-config-docs.sh
@@ -7,7 +7,7 @@ function generate_docs() {
   DOCSPATH=./docs/services/$SERVICE
   echo -en "Creating docs for \e[96m$SERVICE\e[0m... "
   mkdir -p "$DOCSPATH"
-  go run ./cli docs -f markdown "$SERVICE" > "$DOCSPATH"/config.md
+  go run ./cmd/shoutrrr docs -f markdown "$SERVICE" > "$DOCSPATH"/config.md
   if [ $? ]; then
     echo -e "Done!"
   fi
