+++ shoutrrr/cmd/generate/generate.go
