+++ shoutrrr/main.go
@@ -1,15 +1,16 @@
 package main
 
 import (
-	cli "github.com/containrrr/shoutrrr/cli/cmd"
-	"github.com/containrrr/shoutrrr/cli/cmd/docs"
-	"github.com/containrrr/shoutrrr/cli/cmd/generate"
-	"github.com/containrrr/shoutrrr/cli/cmd/send"
-	"github.com/containrrr/shoutrrr/cli/cmd/verify"
+	"os"
+
 	"github.com/containrrr/shoutrrr/internal/meta"
+	cli "github.com/containrrr/shoutrrr/shoutrrr/cmd"
+	"github.com/containrrr/shoutrrr/shoutrrr/cmd/docs"
+	"github.com/containrrr/shoutrrr/shoutrrr/cmd/generate"
+	"github.com/containrrr/shoutrrr/shoutrrr/cmd/send"
+	"github.com/containrrr/shoutrrr/shoutrrr/cmd/verify"
 	"github.com/spf13/cobra"
 	"github.com/spf13/viper"
-	"os"
 )
 
 var cmd = &cobra.Command{
